+++ changelog.md
@@ -6,6 +6,8 @@
 
 - Added debug mode. Enabling it will log the payload and current state whenever a method is triggered.
 
+- Added `Tamnoon.MethodManager.diff/2`.
+
 - Fixed bugs where unrelated HTML classes interfere with Tamnoon classes.
 
 - Added support for setting initial_state as a function, allowing the initial state to be recomputed for every client.